Fix handling of booleans
Booleans are currently handled very poorly.

Especially when the context contains real `true` or `false` booleans.

The Amplitude UI often shows `True` and `False` as auto completions.

Since this SDK uses `strval` to convert a boolean to a string value,
a boolean `true` becomes string `1`. And boolean `false` becomes string `0`.

But if the flag's condition uses string `True` or `False`, it won't match.

To solve this, we add more test cases, and refactor the `matchesIs` in the EvaluationEngine.